+++ app/room/components/room/roomAccess.tsx
@@ -1,0 +1,89 @@
+"use client";
+
+import { validateToken } from "@/app/actions/roomActions";
+import {
+  getRoomState,
+  validatedRoomPasswordAndUsername,
+} from "@/app/actions/userActions";
+import "@livekit/components-styles";
+import { useEffect, useState, FormEvent } from "react";
+import React from "react";
+import RoomAccessForm from "./roomAccessForm";
+
+interface RoomProps {
+  setToken: (token: string) => void;
+  permissionToken: string;
+}
+
+const RoomAccess = ({ setToken, permissionToken }: RoomProps) => {
+  const [isRoomPublic, setIsRoomPublic] = useState<boolean>(true);
+
+  const [roomExists, setRoomExists] = useState<boolean>(true);
+  const [roomId, setRoomId] = useState("");
+
+  useEffect(() => {
+    const fetchRoomState = async () => {
+      const { valid, room } = await validateToken(permissionToken);
+      if (!valid) {
+        setRoomExists(false);
+        return;
+      }
+
+      setRoomId(room!);
+      const roomState = await getRoomState(room!);
+      if (roomState.valid) {
+        setIsRoomPublic(roomState.roomPublic!);
+      } else {
+        setRoomExists(false);
+      }
+    };
+
+    fetchRoomState();
+  }, [permissionToken, roomId]);
+
+  const submitUsernameAndPassword = async (
+    userId: string,
+    password: string
+  ) => {
+    let { message, token } = await validatedRoomPasswordAndUsername(
+      roomId,
+      userId,
+      password,
+      permissionToken
+    );
+    if (token) {
+      setToken(token);
+    }
+    if (message !== "") {
+      alert(message);
+    }
+  };
+
+  if (!roomExists) {
+    return (
+      <>
+        <h1 className="font-bold text-xl">Sorry, couldn't find the room</h1>
+        <p>Couldn't find the room you are looking for</p>
+      </>
+    );
+  }
+
+  // Loading
+  if (roomId === "") {
+    return (
+      <div className="flex flex-col gap-2 m-5">
+        <div className="skeleton h-[48px] w-[250px]"></div>
+        <div className="skeleton h-[48px] w-[250px]"></div>
+        <div className="skeleton h-[48px] w-[62px]"></div>
+      </div>
+    );
+  }
+  return (
+    <RoomAccessForm
+      submitUsernameAndPassword={submitUsernameAndPassword}
+      isRoomPublic={isRoomPublic}
+    />
+  );
+};
+
+export default RoomAccess;
